Skip Nix installation step if installed already
On Travis it should still install Nix. But locally there is no need.
This shaves off a dozen seconds or so from running ./build-nix.sh.

+++ build-nix.sh
@@ -3,51 +3,52 @@
 # Stop if any command fails.
 set -e
 
-printf ':: Downloading and verifying Nix binary ...\n\n'
+nix_profile_sh="$HOME/.nix-profile/etc/profile.d/nix.sh"
 
-mkdir -p downloads
+# Download and install Nix if it isn't already present on the system.
+if [ ! -f $nix_profile_sh ]; then
 
-nixv="nix-2.0.2-x86_64-linux"
+  printf ':: Downloading and verifying Nix binary ...\n\n'
 
-wget --no-clobber --directory-prefix=downloads 'https://nixos.org/nix/install'
-wget --no-clobber --directory-prefix=downloads 'https://nixos.org/nix/install.sig'
-wget --no-clobber --directory-prefix=downloads "https://nixos.org/releases/nix/nix-2.0.2/$nixv.tar.bz2"
+  nixv="nix-2.0.2-x86_64-linux"
+  mkdir -p downloads
+  wget --no-clobber --directory-prefix=downloads 'https://nixos.org/nix/install'
+  wget --no-clobber --directory-prefix=downloads 'https://nixos.org/nix/install.sig'
+  wget --no-clobber --directory-prefix=downloads "https://nixos.org/releases/nix/nix-2.0.2/$nixv.tar.bz2"
 
-# Stored locally to avoid hitting the network every time; `gpg --import` will
-# still try to download the key even if it has it locally. The key fingerprint
-# is B541 D553 0127 0E0B CF15 CA5D 8170 B472 6D71 98DE.
-gpg --import nix-signing-key.gpg
-gpg --verify downloads/install.sig
+  # Stored locally to avoid hitting the network every time; `gpg --import` will
+  # still try to download the key even if it has it locally. The key fingerprint
+  # is B541 D553 0127 0E0B CF15 CA5D 8170 B472 6D71 98DE.
+  gpg --import nix-signing-key.gpg
+  gpg --verify downloads/install.sig
 
-printf ':: Adding build users\n\n'
+  printf ':: Adding build users\n\n'
 
-# For some reason things started failing when running this script in an Ubuntu
-# container, because /usr/sbin (which contains groupadd and useradd) was not in
-# the path. So add it to the path.
-export PATH="$PATH:/usr/sbin"
+  # For some reason things started failing when running this script in an Ubuntu
+  # container, because /usr/sbin (which contains groupadd and useradd) was not
+  # in the path. So add it to the path.
+  export PATH="$PATH:/usr/sbin"
 
-groupadd --force --system nixbld
-for i in `seq -w 1 10`; do
-  useradd -g nixbld -G nixbld              \
-          -d /var/empty -s `which nologin` \
-          -c "Nix build user $i" --system  \
-          nixbuilder$i || true;
-done
+  groupadd --force --system nixbld
+  for i in `seq -w 1 10`; do
+    useradd -g nixbld -G nixbld              \
+            -d /var/empty -s `which nologin` \
+            -c "Nix build user $i" --system  \
+            nixbuilder$i || true;
+  done
 
-printf ':: Extracting and installing Nix ...\n\n'
+  printf ':: Extracting and installing Nix ...\n\n'
 
-mkdir -p /tmp/nix-unpack
-tar -xjf "downloads/$nixv.tar.bz2" -C /tmp/nix-unpack
-/tmp/nix-unpack/$nixv/install
-rm -fr /tmp/nix-unpack
+  mkdir -p /tmp/nix-unpack
+  tar -xjf "downloads/$nixv.tar.bz2" -C /tmp/nix-unpack
+  /tmp/nix-unpack/$nixv/install
+  rm -fr /tmp/nix-unpack
 
-source "$HOME/.nix-profile/etc/profile.d/nix.sh"
+fi # End of Nix install.
 
-printf ':: Building Nginx and acme-client ...\n\n'
+source $nix_profile_sh
 
-nix build -f channel:nixos-18.03 nginx acme-client brotli
-
-printf ':: Building image ...\n\n'
+printf ':: Building nginx, acme-client, and image ...\n\n'
 
 nix build
 image_path=$(nix path-info -f default.nix)
